incorporate modules to core
