+++ core/helpers/index.ts
@@ -4,6 +4,8 @@ import { formatCategoryLink } from '@vue-storefront/core/modules/url/helpers'
 import Vue from 'vue'
 import config from 'config'
 import { sha3_224 } from 'js-sha3'
+import { unicodeAlpha, unicodeAlphaNum } from './validators'
+import store from '@vue-storefront/core/store'
 
 export const processURLAddress = (url: string = '') => {
   if (url.startsWith('/')) return `${config.api.url}${url}`
@@ -200,19 +202,35 @@ export const serial = async promises => {
   return results
 }
 
-export const isBottomVisible = () => {
-  if (isServer) {
-    return false
-  }
-  const scrollY = window.scrollY
-  const visible = window.innerHeight
-  const pageHeight = document.documentElement.scrollHeight
-  const bottomOfPage = visible + scrollY >= pageHeight
-
-  return bottomOfPage || pageHeight < visible
-}
-
 // helper to calcuate the hash of the shopping cart
 export const calcItemsHmac = (items, token) => {
   return sha3_224(JSON.stringify({ items, token: token }))
+}
+
+export function extendStore (moduleName: string | string[], module: any) {
+  const merge = function (object: any = {}, source: any) {
+    for (let key in source) {
+      if (typeof source[key] === 'object') {
+        object[key] = merge(object[key], source[key])
+      } else {
+        object[key] = source[key]
+      }
+    }
+    return object
+  };
+  moduleName = Array.isArray(moduleName) ? moduleName : [moduleName]
+  const originalModule: any = moduleName.reduce(
+    (state: any, moduleName: string) => state._children[moduleName],
+    (store as any)._modules.root
+  )
+  const rawModule: any = merge({}, originalModule._rawModule)
+  const extendedModule: any = merge(rawModule, module)
+
+  store.unregisterModule(moduleName)
+  store.registerModule(moduleName, extendedModule)
+}
+
+export {
+  unicodeAlpha,
+  unicodeAlphaNum
 }
