+++ core/mixins/onBottomScroll.js
@@ -1,4 +1,16 @@
-import { isBottomVisible } from '@vue-storefront/core/helpers'
+import { isServer } from '@vue-storefront/core/helpers'
+
+const isBottomVisible = () => {
+  if (isServer) {
+    return false
+  }
+  const scrollY = window.scrollY
+  const visible = window.innerHeight
+  const pageHeight = document.documentElement.scrollHeight
+  const bottomOfPage = visible + scrollY >= pageHeight
+
+  return bottomOfPage || pageHeight < visible
+}
 
 /**
  * By implementing this mixin add "onBottomScroll" mthod in component.
