+++ core/modules/catalog-next/index.ts
@@ -1,5 +1,5 @@
 import { categoryModule } from './store/category'
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 
 export const CatalogNextModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
   store.registerModule('category-next', categoryModule)
