+++ core/modules/compare/index.ts
@@ -1,7 +1,7 @@
 
 import { compareStore } from './store'
 import { plugin } from './store/plugin'
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
 
 export const cacheStorage = StorageManager.init('compare')
