+++ core/modules/mailer/index.ts
@@ -1,4 +1,4 @@
-import { StorefrontModule } from '@vue-storefront/module'
+import { StorefrontModule } from '@vue-storefront/core/lib/modules'
 import { mailerStore } from './store'
 
 export const MailerModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
