+++ core/modules/newsletter/index.ts
@@ -1,5 +1,5 @@
 import { newsletterStore } from './store'
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
 
 export const cacheStorage = StorageManager.init('newsletter')
