+++ core/modules/notification/index.ts
@@ -1,5 +1,5 @@
 import { notificationStore } from './store'
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 
 export const NotificationModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
   store.registerModule('notification', notificationStore)
