+++ core/modules/recently-viewed/index.ts
@@ -1,7 +1,7 @@
 import { recentlyViewedStore } from './store'
 import { plugin } from './store/plugin'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 import { isServer } from '@vue-storefront/core/helpers'
 
 export const cacheStorage = StorageManager.init('recently-viewed')
