+++ core/modules/review/index.ts
@@ -1,4 +1,4 @@
-import { StorefrontModule } from '@vue-storefront/module'
+import { StorefrontModule } from '@vue-storefront/core/lib/modules'
 import { reviewStore } from './store'
 
 export const ReviewModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
