+++ core/modules/url/index.ts
@@ -1,5 +1,5 @@
 import { urlStore } from './store'
-import { StorefrontModule } from '@vue-storefront/module'
+import { StorefrontModule } from '@vue-storefront/core/lib/modules'
 import { beforeEachGuard } from './router/beforeEach'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
