+++ core/modules/wishlist/components/AddToWishlist.ts
@@ -1,7 +1,7 @@
 import Product from '@vue-storefront/core/modules/catalog/types/Product'
 import { WishlistModule } from '../'
 import wishlistMountedMixin from '@vue-storefront/core/modules/wishlist/mixins/wishlistMountedMixin'
-import { registerModule } from '@vue-storefront/module';
+import { registerModule } from '@vue-storefront/core/lib/modules';
 
 export const AddToWishlist = {
   name: 'AddToWishlist',
