+++ core/modules/wishlist/index.ts
@@ -1,4 +1,4 @@
-import { StorefrontModule } from '@vue-storefront/module'
+import { StorefrontModule } from '@vue-storefront/core/lib/modules'
 import { wishlistStore } from './store'
 import { plugin } from './store/plugin'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
