+++ core/package.json
@@ -8,7 +8,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@vue-storefront/module": "^0.0.9",
+    "@vue-storefront/core/lib/modules": "^0.0.9",
     "bodybuilder": "2.2.13",
     "config": "^1.30.0",
     "express": "^4.14.0",
