+++ package.json
@@ -42,8 +42,7 @@
     "test:e2e": "cypress open",
     "test:e2e:ci": "cypress run",
     "lint": "eslint --ext .js,.vue,.ts core src --fix",
-    "lerna": "lerna",
-    "pkg:module:build": "cd ./packages/module && yarn build"
+    "lerna": "lerna"
   },
   "pre-commit": [ "lint" ],
   "dependencies": {
