--- packages/module/.gitignore
@@ -1,2 +1,0 @@
-./*.js
-./*.d.ts
