+++ src/modules/hotjar/index.ts
@@ -1,4 +1,4 @@
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 import { isServer } from '@vue-storefront/core/helpers'
 
 const hotjarStore = {
