+++ src/modules/instant-checkout/index.ts
@@ -1,4 +1,4 @@
-import { StorefrontModule } from '@vue-storefront/module';
+import { StorefrontModule } from '@vue-storefront/core/lib/modules';
 
 export const InstantCheckoutModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
 }
